+++ golang/app/cmd/MSword_counter.bash
@@ -1,4 +1,4 @@
-#!/bin/bash
+#!/app/bash
 
 set -ue
