+++ golang/app/cmd/fetch_latest_file.bash
