--- golang/fetchFile/testdata/input.sql
@@ -1,8 +1,0 @@
-USE glory_test;
-
--- ここに入力値を追加
-INSERT INTO author VALUE (0, 'suzuki', 'B4');
-INSERT INTO author VALUE (0, 'takahashi', 'B4');
-
-INSERT INTO thesis_history VALUE (0, 1, 120, '2020-11-15 2:34:00', '2020-11-15 3:00:00');
-INSERT INTO thesis_history VALUE (0, 2, 140, '2020-11-15 2:44:00', '2020-11-15 3:00:00');
