--- golang/fetchFile/testdata/input2.sql
@@ -1,5 +1,0 @@
-USE glory_test;
-
--- ここに入力値を追加
-INSERT INTO author VALUE (0, 'suzuki', 'B4');
-INSERT INTO author VALUE (0, 'takahashi', 'B4');
