--- golang/fetchFile/testdata/teardown.sql
@@ -1,6 +1,0 @@
-USE glory_test;
-
-set foreign_key_checks = 0;
-TRUNCATE TABLE thesis_history;
-TRUNCATE TABLE author;
-set foreign_key_checks = 1;
