+++ golang/go.mod
@@ -4,12 +4,10 @@ go 1.15
 
 require (
 	github.com/Songmu/flextime v0.1.0
+	github.com/gin-contrib/multitemplate v0.0.0-20220829131020-8c2a8441bc2b
+	github.com/gin-gonic/gin v1.8.1
 	github.com/go-sql-driver/mysql v1.6.0
-	github.com/golang/protobuf v1.4.3 // indirect
-	github.com/google/go-cmp v0.5.2
-	github.com/jmoiron/sqlx v1.3.5 // indirect
-	github.com/joho/godotenv v1.4.0 // indirect
-	golang.org/x/oauth2 v0.0.0-20200902213428-5d25da1a8d43
-	golang.org/x/sys v0.0.0-20201130072748-111129e158e2 // indirect
-	google.golang.org/api v0.35.0
+	github.com/jmoiron/sqlx v1.3.5
+	golang.org/x/oauth2 v0.3.0 // indirect
+	google.golang.org/api v0.105.0 // indirect
 )
