--- golang/web/go.mod
@@ -1,23 +1,0 @@
-module github.com/westlab/glory/web
-
-go 1.15
-
-require (
-	github.com/Songmu/flextime v0.1.0
-	github.com/gin-contrib/multitemplate v0.0.0-20200916052041-666a7309d230
-	github.com/gin-gonic/gin v1.6.3
-	github.com/go-playground/validator/v10 v10.4.1 // indirect
-	github.com/go-sql-driver/mysql v1.6.0
-	github.com/google/go-cmp v0.5.2
-	github.com/jmoiron/sqlx v1.3.5 // indirect
-	github.com/json-iterator/go v1.1.10 // indirect
-	github.com/modern-go/concurrent v0.0.0-20180306012644-bacd9c7ef1dd // indirect
-	github.com/modern-go/reflect2 v1.0.1 // indirect
-	github.com/ugorji/go v1.2.0 // indirect
-	github.com/westlab/glory v0.1.0
-	golang.org/x/crypto v0.0.0-20201124201722-c8d3bf9c5392 // indirect
-	golang.org/x/sys v0.3.0 // indirect
-	gopkg.in/yaml.v2 v2.4.0 // indirect
-)
-
-replace github.com/westlab/glory v0.1.0 => ../
